Depost and Withdraw method done.

+++ app/src/main/java/com/comp3350/webudget/persistence/IWalletDatabase.java
@@ -9,6 +9,6 @@ public interface IWalletDatabase {
 
     public int insertWallet(String username); //returns the ID of the wallet
     public Wallet getWallet(int id) throws WalletException;
-    public void deposit(int walletID, int amount) throws WalletException;
-    public void withdraw(int walletID, int amount) throws WalletException;
+    public void deposit(int walletID, double amount) throws WalletException;
+    public void withdraw(int walletID, double amount) throws WalletException;
 }
